requirements.txt
fire==0.6.0
deepl==1.18.0
python-dotenv==1.0.0
